Add job to verify upstream global-jjb patches
Add job that can be used by downstream projects to report upstream
to global-jjb patches when a proposed patch breaks their project.

This job sets -1 CR only and does not block the upstream patch as
there are cases where we may want to break backwards compatibility
on purpose. This job's purpose is to give us a clear view when it
does happen.

Change-Id: If0dda88ff24da7fd171484c37855f33b05f36553
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>