Add option to stage to Maven Central
Provide an option to also stage to OSSRH in preparation to release
to Maven Central.

In addition to printing the staging repo-id also print the URL to
the staging repo for convenience in stage-repo.txt.

NOTE: This patch introduces a breaking change to
      lf-infra-maven-parameters to introduce a new boolean parameter
      mvn-central. Job templates using this macro will need to add
      'mvn-central: false' to their list of passed parameters.

Issue: https://jira.opendaylight.org/browse/RELENG-28
Change-Id: I0628cb5c7185e2143e448ceb4fa4752384581a5e
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>