Make maven-merge jobs run daily
Back when OpenDaylight was on Rackspace we were trying to cut back
on how often jobs run due to the infra being unstable and at the
time thought the maven-integration job and maven-merge jobs were
part of the problem. Since we moved to Vexxhost infra has been
significantly more stable and not as big a concern. Let's run
the maven-merge jobs on a daily basis so that project artifacts
have more chances to pass before the 3 week SNAPSHOT clearing
cron that's run in Nexus.

Change-Id: I7ac17b3b4b8729d0f143fb15c71143c3de1ac461
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>