Update patch set 2
Patch Set 2:

Didn't added it in any other macro (neither in the Github CLM) becuase is not a requirement there yet. 

That would be changing too many parts of the code where i need to add this and is not required yet. 

Only CLM gerrit needs it for now.

Patch-set: 2