Allow maven-clm job to specify AppID namespace
This allows maven-clm jobs to insert a project namespace into the
AppID field in Nexus IQ in cases where a Nexus IQ server might be
shared by multiple projects and name collision might be an issue.

This patch inserts an optional nexus-iq-namespace parameter which
by default is blank. This should not break backwards compatibility
while allowing the new namespace feature to work. We recommend when
using the namespace to add a trailing - to the value. Eg. 'odl-',
this is to make the namespace look nice for example "odl-aaa" will be
the result of namespace odl-, and project is aaa.

Issue: RELENG-1307
Change-Id: If4313ba00da67541d337b70f7946ad6f21fee037
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>