Update patch set 1
Patch Set 1:

> (1 comment)
 > 
 > I like this patch, but I wonder if we can start making this
 > automatic so that the teams do not have to worry about one extra
 > variable.

This is possible with this patch. This is the method I would do it:

https://git.opendaylight.org/gerrit/#/c/77031/1/jjb/releng-templates-java.yaml

The reason I'm implementing the maven-clm job this way is so that the job-template changes does not break existing users of the job making it backwards compatible. If we were to utilize defaults.yaml it would be a breaking change and force all global-jjb users to now edit their defaults.yaml regardless of if they care about namespaces or not to insert a blank " namespace: '' " parameter.

Patch-set: 1