Update patch set 2
Patch Set 2:

> I added the file existence check so I won't see 'grep
 > /etc/default/sysstat: No such file or directory' in the log. If the
 > file does exist, the code works exactly as before. I know that
 > 'service' is available on newer versions of Ubuntu but I don't how
 > far back that goes and I don't know what versions of Ubuntu this
 > script needs to support. Either way works for Hyp but I am not the
 > only user.

If sysstat is not available, it would be good to know that by a message in the logs than trying to figure out what is happening when the logs are missing. with that check in place that would go silent which is why I was recommending we should probably check the service status.

Patch-set: 2