Automate ODL tagging and artifact release
This will be adapted to work for ONAP/Other projects
once we are okay with this.

projects that want self-serve releases will create a
tagname.yaml file in the releases/ directory of their git repo.

Here is the change I have been testing in the sandbox
https://git.opendaylight.org/gerrit/#/c/80613/

example:
$ cat releases/1.0.0.yaml
---
artifact_type: 'maven'
version: '1.0.0'
project: 'zzz-test-release'
log_dir: 'zzz-test-release-maven-stage-master/17/'
maven_central_url: 'oss.sonatype.org'

We might be able to remove some of these fields,
but this is good for a first review.

Some comments from Thanh that look correct to implement next

So these release jobs are currently for ODL Maven style jobs.
Some forward thinking...
Releasing "Docker" for example will be handled differently. I'm
wondering if we want to create parent jobs to handle the different
cases?
Eg.

  - release-tag-verify  # verifies release tags but does not push
    - release-maven-verify  # verifies the release files availablilty
but does not release
  - release-tag-merge  # handles tags
    - release-maven-merge  # handles maven artifacts
    - release docker-merge  # ... future handles docker artifacts
In this case the release-tag-* jobs are the parent jobs, they will use
the "parameterized trigger" to trigger addidtional jobs as necessary
depending on project as blocking jobs before releasing the tags.

Issue: RELENG-1657

Change-Id: I02dc935a84425d194cdd8eaabe5b43c5920ce319
Signed-off-by: Aric Gardner <agardner@linuxfoundation.org>