Automate ODL tagging and artifact release
This will be adapted to work for ONAP/Other projects
once we are okay with this.

projects that want self-serve releases will create a
tagname.yaml file in the releases/ directory of their git repo.

example:
$ cat releases/4.0.0.yaml
---
artifact_type: 'maven'
version: '4.0.0'
project: 'odlparent'
log_dir: 'odlparent-maven-release-master/11/'
maven_central_url: 'oss.sonatype.org'

We might be able to remove some of these fields,
but this is good for a first review.

Issue: RELENG-1657

Change-Id: I02dc935a84425d194cdd8eaabe5b43c5920ce319
Signed-off-by: Aric Gardner <agardner@linuxfoundation.org>