Update patch set 1
Patch Set 1:

(2 comments)

The hyperledger project needs the ability to allocate '/w' in tmpfs and the ability to specify the swap size. If we are to use the init scripts from global-jjb (a good thing) we needs to allow these customization. I have proposed 'a' way to do this, but it is not the only way.

Patch-set: 1
Label: Code-Review=0