Install lftools 0.17.1 for deploy-maven-files.sh
An undetected error when adapting "deploy maven-file" command to
pure python is that Nexus does not allow manual uploads to SNAPSHOT
repos. This may make the very idea of adapting "deploy maven-file"
to use the REST API impractical. For now, this change will allow
jobs that currently rely on this shell script to work as intended.

Change-Id: I655494b7754da95eb63d3785c1cba7a3d9191ef6
Signed-off-by: Eric Ball <eball@linuxfoundation.org>