Update patch set 1
Patch Set 1:

> I need to understand why this is being set in the script here ...
 > this may impact all other projects?

As stated in the commit message, newer versions of lftools are using "requests" to access the Nexus REST API. This is not compatible with pushing to SNAPSHOT repos. This affects several jobs in Akraino, and some of the netvirt, opflex, and packaging jobs in ODL. Since none of the jobs were ever modified to use the updated version of this command (which uses different arguments), setting the version back to a known-good version should fix them without causing any issues.

Patch-set: 1