Update patch set 1
Patch Set 1:

> Ok, after discussing with Anil, I'm going to wait until we can
 > discuss this further with Thanh, and look at less hacky
 > alternatives. I'll leave this up for reference, but it will not be
 > merged until further discussion has occurred. I'm going to look
 > into rolling back global-jjb in the Akraino for the time being as a
 > possible temporary solution to the failing builds.

I agree, we should take Thanh's input on this, since we are assuming that lftools is broken with REST API, and look at one of the other ways. 

We can release a new version of global-jjb with this workaround, if everyone is one the same page and agrees this is the only solution.


Other alternatives:
1. Rollback of lftools to a working version (v0.17) in global-jjb.
2. Revert the change in lftools and do a new lftools release to fix it in lftools.
3. If projects are impacted with latest version of global-jjb with the upgrade, we should consider moving them back to a previous version.

Patch-set: 1