Update rtd verify job to check release notes format
The rtd verify job presently does not fail if there are
formatting issues with release notes.

To fix this we can call  `reno report` through tox which does
a lint check and converts the release notes in text files.
The command fails the verify job on any incorrect formatting.

Issue: RELENG-1863
Change-Id: I9d47ae9b57e6f5fd9858a44cf05f6fe81d3278c5
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>