Update patch set 1
Patch Set 1:

I think the only change that needs to happen is here:

https://github.com/opendaylight/releng-builder/blob/master/jjb/lf-infra/lf-infra-jobs.yaml#L112-L128


To add:

     gerrit_trigger_file_paths:
      - compare-type: REG_EXP
        pattern: '.*\.css'
      - compare-type: REG_EXP
        pattern: '.*\.html'
      - compare-type: REG_EXP
        pattern: '.*\.rst'
      - compare-type: REG_EXP
        pattern: '.*\/conf.py'
      - compare-type: REG_EXP
        pattern: 'releasenotes/.*'


under the rtd-jobs to cause it to trigger the rtd-verify job to trigger when a change in releasenotes/.* directory is detected.

I think the change in this patch is overcomplicating something that the existing rtd-verify job could already easily handle.

Patch-set: 1