INFO.yaml check vote for global-jjb
Global job per gerrit, that triggers on changes
to INFO.yaml file

Problems: for opnfv we dont have
centos7-builder-2c-1g avaliable
so the top of the shell script that
installs lftools will not work.

This job should merge the change
It does not yep attempt to do this.
We will need to give
Non-Interactive Users merge
or perhaps specifically only the jenkins
user.

Example of Job configuration:

- project:
    name: info-vote
    build-node: centos7-builder-2c-1g
    jobs:
      - 'info-vote-verify'

Change-Id: Ib386bd2ea51be113ca59610bd3281b9ad6923bce
Signed-off-by: Aric Gardner <agardner@linuxfoundation.org>