INFO.yaml check vote for global-jjb
Global job per gerrit, that triggers on changes
to INFO.yaml file

And auto merged if a majority of committers
and tsc members have voted.

We will need to give
Non-Interactive Users merge
or perhaps specifically only the jenkins
user.

Job will only trigger when a user votes +2
cludgy but it works.

Example of Job configuration:

- project:
    name: info-vote
    build-node: centos7-builder-2c-1g
    jobs:
      - 'info-vote-verify'

Change-Id: Ib386bd2ea51be113ca59610bd3281b9ad6923bce
Signed-off-by: Aric Gardner <agardner@linuxfoundation.org>