DRAFT: Add tox-verify job to the default ci group
This ensures our projects are running linters against the prjoect.

!! Do not merge this one yet. !!
Before then we should ensure all ci-management projects have some form of
tox.ini / .coafile linting in place so that when we enable this job it
doesn't cause them a headache of having to fix all the linting right away.

Change-Id: Idd5e8f44408b16ef5a2051460cf5f3bb687881a7
Signed-off-by: Bengt Thuree <bthuree@linuxfoundation.org>