Fixes for info votes
gerrit trigger is not working in opnfv.
we had this happen to some of the read the docs jobs...
not sure what the fix is.

Not sure were to test to approval category stuff.

That said I think the mistake is that the value is an int.

            approval-category: 'WHATEVER'
            approval-value: 2

I dont actually want the script to exit on first error, need to output
the vote verdict.

not sure why I thought that JENKINS_SSH_CREDENTIAL
was the correct var for the ssh user.
switched to $USER

Change-Id: I6714361423948eb1764e9c18da699dafd4cd24a9
Signed-off-by: Aric Gardner <agardner@linuxfoundation.org>