Fixes for info votes
gerrit trigger is not working in opnfv.
we had this happen to some of the read the docs jobs...
not sure what the fix is.

For some reason the comment added approval category stuff
did not show up in the job config.
Switching the order of the values, maybe?
or like we cant really handle both cases and I need to make this
configurable.

I dont actually want the script to exit on first error, need to output
the vote verdict.

not sure why I thought that JENKINS_SSH_CREDENTIAL
was the correct var for the ssh user.
switched to $USER

Change-Id: I6714361423948eb1764e9c18da699dafd4cd24a9
Signed-off-by: Aric Gardner <agardner@linuxfoundation.org>