Update patch set 1
Patch Set 1: Code-Review-1

I'd recommend not to remove the debug output from the console logs. Its always been very useful in determining what files are getting archived. If we don't want this to be in console logs, please capture that output into a diff ${file-name} that can be archived, instead of having it in the console logs.

Patch-set: 1
Label: Code-Review=-1