Update patch set 4
Patch Set 4:

Decided not to archive the diff that was generated. The changes I have substantially reduced the output but with the diff output in the log the necessary debug info has been retained. If the diffs end up being too much info in the log we can archive the diffs at a later time.

Tested on a couple of fabric builds and a couple of ci-management builds.

Patch-set: 4