Make stripping of "SNAPSHOT" in poms optional
When maven-versions-plugin is set to "true", the job will
use the versions plugin to update the versions in the pom.xml
files and will not need this step in maven-patch-release.sh to
happen.
This change DOES NOT affect current jobs which do not use the
maven versions plugin as this variable is set to false by default.
Update maven-versions-plugin variable in the parameters to be able
to parse the value in the scripts.

Change-Id: Ic25197012fd0f23a859159bb86369feb8a02b9c2
Signed-off-by: Jessica Wagantall <jwagantall@linuxfoundation.org>