Add submodule-disable for submodule operations
Some Akraino projects using git scm have a large number of submodules
which are not required during most CI/CD operations (e.g. we don't
want to fetch GBs of 3rd party repositories when linting the code).

Add submodule-disable parameter for both lf-infra-gerrit-scm
and lf-infra-github-scm, allowing us to disable submodule fetching
where needed.

Jenkins git plugin support for this was added in [1].

This change is very similar and mostly based on the addition of the
submodule-timeout parameter in commit 9fdb611f1.

[1] https://issues.jenkins-ci.org/browse/JENKINS-6658

Change-Id: Ia99ba7f353a39e55d7bec57778f447f94402b5e6
Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>