Update patch set 6
Patch Set 6: Code-Review-1

(14 comments)

> This was motivated by comments on https://gerrit.oran-osc.org/r/#/c/78/.
 > 
 > Andrew and other JJB experts, please tell me how I could reduce the
 > amount of repeated code.  My new macros are 96% identical to your
 > existing macros which seems like a real opportunity for
 > improvement.

Instead of completely new job templates, we could instead move all the parts that different / optional into conditional build-steps. Default them to the base job design and then just enable the steps (and disable other portions) based on the need of the job.

Alternatively, we break out the parts that are primarily the same into new anchors and include those in both sets and then just layer on the final differences

 > I don't know how to do release notes here :/

it's done as follows:

tox -e reno my_change_note_title

Then edit the resulting file and add that to the change.

Patch-set: 6
Label: Code-Review=-1