Remove maven-versions-plugin-set-version var
This is a variable that does not need to be defined by the users
of the jobs. The version needed in this builder step is inherited
from versions.properties as "release_version" and it is fixed as
that.
This also helps teams to not having to define this version in 2
places and just reply on version.properties.
This upgrade only affects projects using maven-versions-plugin
to update their versions (ONAP).

Change-Id: I6393291eb555ed0b8f5220264688193c1af9fd5f
Signed-off-by: Jessica Wagantall <jwagantall@linuxfoundation.org>