Update patch set 10
Patch Set 10:

Actually.. it makes sense to me to not run daily the merge job and save resources.. but i want people to comment if im right cuz this is a macro that others are using. 

For now, is not a big deal to have it daily... 

Team, can you guys please comment?

Patch-set: 10