Update patch set 10
Patch Set 10:

In case of ONAP for example, we create a new snapshot after every merged commit. The snapshot version can be overwritten if the team decides to re-push it and this will preserve the snapshot around as long as it is "refreshed". 

We do encourage teams to not depend on snapshots and to make releases instead so that they dont run into issues if their dependencies are on old versions of a snapshot that might get removed if time expires. 

I think the teams can still make as many pushes of the same snapshot as they need but is not a desired practice to have these dependencies.

Patch-set: 10
Label: Code-Review=0