Remove maven-versions-plugin-set-version var
Clean up newly added maven-docker-stage macro that missed a recent change.
This is a variable that does not need to be defined by the users of the jobs.
The version needed in this builder step is inherited from versions.properties
as "release_version". This also helps teams not having to define this version
in 2 places and just rely on version.properties.

Change-Id: I41083ed784982b904e12edf52d1f8e73a249b13d
Signed-off-by: Lott, Christopher (cl778h) <cl778h@att.com>