Update patch set 1
Patch Set 1:

This appears to be a breaking change. I attempted to upgrade the Acumos ci-management to the latest global-jjb which includes this change, and now validation fails.  Do all users of this level global-jjb need to add a value for submodule-disable to their defaults.yaml file?  What is the recommended new value, true or false?  Is it possible to add a reasonable default value that clients can override, so this isn't a braking change?  Please advise, thanks.

Patch-set: 1