Update patch set 1
Patch Set 1:

I think I understand now why my Acumos-JJB breaks on this change to global-jjb macro lf-infra-gerrit-scm.  The change is in the doc, pasted below.  I find it confusing that this macro's required parameters show default values. It seems to me that either a parameter is required (no default), or the parameter is optional (and has a default).  Please clarify the doc.  

I think I see how to fix my macros and templates that use lf-infra-gerrit-scm.  But maybe I'm not the only one who reused this?  Could this macro possibly be extended so the new parameter is optional and defaults to false?

    lf-infra-gerrit-scm
    -------------------

    Basic SCM configuration for Gerrit based projects.

    :Required parameters:

    :submodule-timeout: Timeout (in minutes) for checkout operation.
        (default: 10)
    :submodule-disable: Disable submodule checkout operation.
        (default: false)

Patch-set: 1