Update patch set 1
Patch Set 1:

> I think I understand now why my Acumos-JJB breaks on this change to
 > global-jjb macro lf-infra-gerrit-scm.  The change is in the doc,
 > pasted below.  I find it confusing that this macro's required
 > parameters show default values. It seems to me that either a
 > parameter is required (no default), or the parameter is optional
 > (and has a default).  Please clarify the doc.
 > 
 > I think I see how to fix my macros and templates that use
 > lf-infra-gerrit-scm.  But maybe I'm not the only one who reused
 > this?  Could this macro possibly be extended so the new parameter
 > is optional and defaults to false?
 > 
 > lf-infra-gerrit-scm
 > -------------------
 > 
 > Basic SCM configuration for Gerrit based projects.
 > 
 > :Required parameters:
 > 
 > :submodule-timeout: Timeout (in minutes) for checkout operation.
 > (default: 10)
 > :submodule-disable: Disable submodule checkout operation.
 > (default: false)

Hi,
Unfortunately this is not possible with macros - default values can only be used for jobs or job-templates.
As per the release notes for the upgrade procedure [1], the new 'submodule-disable' should be added to all consumers of these 2 macros. The default value (false) is the same with the default value used for this by the Gerrit Jenkins plugin.
I wish we could have implemented this in a way that did not require updating your defaults when bumping global-jjb, however afaik there was no way around it ...

[1] https://docs.releng.linuxfoundation.org/projects/global-jjb/en/latest/release-notes.html

Patch-set: 1