Update patch set 5
Patch Set 5:

> Thanks Thanh - please add the appropriate tag - maybe v0.37.3?

Considering master also includes the submodule-disable change which is a breaking change I don't think a patch release is appropriate here. Unless we want to open a v0.37.x maintenance branch based on the last release and cherry-pick the patch over.

With that said I'll leave it up to LF staff to decide how they want to handle this release.

Patch-set: 5