Update patch set 11
Patch Set 11:

One more question - is there any agreement for where to get the version (tag) string for a docker image in non-maven projects?  In a maven project you just fetch the project.version string; of course not all the world is maven.  This seems really, really important, it seems like it would be chaos if a version number is NOT used in the docker image tag.

Patch-set: 11