Fix builder steps in lf-maven-docker-stage
lf-maven-versions-plugin builder step needs to run before
maven-patch-release.sh as this second script contains a
condition to confirm if the maven vesions plugin was selected
as a way to remove the 'SNAPSHOT' pattern from the pom.xml
files.
lf-maven-docker-stage was based on lf-maven-stage and it
seems that these particular builder steps were switched in
place accidentally.

Change-Id: I30ac4af2a7566cb2c1072412a738f38d4f3318fb
Signed-off-by: Jessica Wagantall <jwagantall@linuxfoundation.org>