Update patch set 17
Patch Set 17:

I like that you're trying to make this very robust, but if a precondition is violated, why isn't it ok to stop there?  If a project chooses git-describe as the docker-tagging mechanism, the project is saying that they also have chosen git-tagging in their regular workflow.  I don't think extra logic is required in your script.  If a script is directed to call git describe, and that answers empty due to absence of tags, fail the build.

Patch-set: 17