Update patch set 17
Patch Set 17:

Is there any value in refactoring so that the docker-tag-choosing scripts limit themselves to only that, namely setting the docker image name (complete with registry, name and tag) for use by later stages? Then add a new docker-build script that relies on finding a docker image name  in an environment variable and invoking docker build. Just wanted to raise the idea.

Patch-set: 17
Label: Code-Review=0