Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> > Patch Set 9:
> > 
> > Still seeing this error with this change:
> > 
> > 09:04:44 + PATH=/usr/local/bin:/usr/bin:/opt/puppetlabs/bin:/usr/local/sbin:/usr/sbin:/home/jenkins/.local/bin:/home/jenkins/bin
> > 09:04:44 + export PATH
> > 09:04:44 + echo '---> logs-deploy.sh'
> > 09:04:44 ---> logs-deploy.sh
> > 09:04:44 + set -euf -o pipefail
> > 09:04:44 + [[ -z https://logs.opendaylight.org ]]
> > 09:04:44 + nexus_url=https://nexus.opendaylight.org
> > 09:04:44 + nexus_path=sandbox/vex-yul-odl-jenkins-2/autorelease-update-submodules-fluorine/1
> > 09:04:44 + pattern_opts=()
> > 09:04:44 /tmp/jenkins3743036912519971292.sh: line 29: pattern_opts[@]: unbound variable
> > 09:04:44 [autorelease-update-submodules-fluorine] $ /bin/bash /tmp/jenkins6532085800486700684.sh
> > 09:04:44 [description-setter] Could not determine description.
> > 09:04:44 Build step 'Execute Scripts' changed build result to UNSTABLE
> 
> It initializes 'pattern_opts' as an empty array, and in the next line it references the array, how can it not be set ?

I now understand 'part' of the problem... note that -x is set when the script is running, and yet we do not set it in the script.... that is because this script is NOT getting executed, it is being sourced. I have always hated the way Jenkins executes scripts (they had to be compatible with windoze). '- shell: !include-raw:' does NOT do what you think... or what we want. I need to do some digging... still don't see how this causes 'unbound variable'.

Patch-set: 9