Update patch set 1
Patch Set 1:

If a stack is stuck on DELETE_FAILED, sometimes we can remove it after clearing up the orphaned nodes or port, however sometimes we cant and require (special power) to raise a ticket with vexx to clean this up. So with the job lftools should fail throwing an error. 
This be dealt in the scripts to exit gracefully, without failing the jobs.

Patch-set: 1
Reviewer: Gerrit User 29 <29@f2a2391d-0403-4fa5-b183-4056999a8a09>
Label: Code-Review=0