Update patch set 1
Patch Set 1:

I did a search on fallocate swap and the first two hits said it was OK to fallocate. I did not read the third, which was the link you provided...sigh. For my image Ubuntu 16.04 swapon did not generate an error or warning. We could check status of swapon and redo with 'dd' if swapon fails. It is not clear that swapon will always complain if there is a problem.

I do not think using 'dd' to create swap at boot time is a viable option. 40sec for 5GB and 130 secs for 10GB just takes too long. The swap file either needs to be allocated in the image or moved out of 'Jenkins Init' phase into the project phase so each project can weigh the costs of creating swap space. There could a lf-infra macro ? or script. Either would have to 'sudo'.

Patch-set: 1