Update patch set 21
Patch Set 21:

Thanks! i think it will be the best given the fact that the file is still a maven type in the system. 

Let me know if you see anything else we could change in my patch.. i am testing this too and see if i spot anything too.

Patch-set: 21