Update patch set 24
Patch Set 24:

> Patch Set 24:
> 
> You all merge your own code!?  Well that's fine by me.  Thanks Jessica!  I'll upgrade our ci-management global-jjb submodule and go try it.

Yes an no. We have NACR in place so the owner/author can't self-vote in a change, but they can do the final submit as long as someone else has given a +2. All of our other projects actually work this way to, folks just don't realize it.

We sometimes give final vote so that the owner can merge it when it seems appropriate. This happens most often with our puppet infrastructure as folks sometimes need to handle things in a specific order, even though puppet is doing the actual work.

Patch-set: 24