Update lf-docker-jobs to enforce macro vars
Enforce the varaibles from "lf-docker-get-container-tag",
"lf-docker-build" and "lf-docker-push" to be carried over
when calling the macros. This avoids the issue of loosing
the value of the variables.
Update the test file to reflect the example better.
Modify scripts to fix variable convention. Remove push regisrty and use
if only after docker-push.sh is called.
Allow docker-build-args to be undefined as it is not mandatory.

Change-Id: Ic3aad7ae0f87b589ce5d57f6848f7e8e75f9b71e
Signed-off-by: Jessica Wagantall <jwagantall@linuxfoundation.org>