Update patch set 3
Patch Set 3: Code-Review-1

The push in the merge job fails because the image is not named correctly when built. Here's an example from sandbox job https://jenkins.o-ran-sc.org/sandbox/job/ci-management-build-ubuntu-c-cpp-docker-merge-master/1/console

16:56:04 ---> docker-push.sh
16:56:04 The push refers to repository [nexus3.o-ran-sc.org:10004/build-ubuntu-c-cpp]
16:56:04 An image does not exist locally with the tag: nexus3.o-ran-sc.org:10004/build-ubuntu-c-cpp
16:56:04 Build step 'Execute shell' marked build as failure

The docker build step must name/tag the image with the complete URL of the registry - host:port/name:tag.  The docker push command accepts the exact complete URL and does the pushing.  One of the things that has always confused me about docker is that the push command does not allow you to name a target - it's not like scp - docker push just works on an image that is present locally.  

That's a long way of saying, please restore the previous code in docker-build.sh that uses the full name in the build step.

Patch-set: 3