Update patch set 10
Patch Set 10: -Code-Review

Thanks for explaining how 1.0 in the file became just tag 1.  Please update the example in the documentation to quote the string, that's just good practice because it should be a string in all cases, never just an integer.

And if you'll be so kind to have the get-tag scripts echo their results, I'll give +1.

Patch-set: 10
Label: Code-Review=0