Update patch set 11
Patch Set 11:

By using a bare tag 1.0 I triggered a Yaml behavior to treat it as a number.  I guess if I had used 1.0.0 like your example, or any leading alpha character, it would not have happened.  So, I guess there is no special need to quote the string in the example in the doc.

Patch-set: 11