Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Thanks Andrew!
> 
> I will need to upgrade the o-ran-sc ci-management area to the latest global-jjb to pick this up.  But, it appears you really (really) prefer doing that task yourself.  Please advise how to proceed on that.

Already upgraded it :D

It's not so much that I really prefer doing that task, it just so happens that I already had an upgrade open when you had made yours. I personally prefer that the upgrades all go against the same topic and have a similar commit message.

The topic I use against all gerrit systems is:

global-jjb_upgrade

And the commit message is usually very short like:

Upgrade to global-jjb v0.38.2

If there happens to be modifications that are happening outside of just the submodule update because of the upgrade, I make note in the commit message body as to why there is more than just a submodule update.

This makes it easy to track down all the updates that have happened. If the update is against a non-released / tagged version then a more detailed commit message is also done with reasons why.

So, if you follow that, then I'll merge your updates too ;) But you might consider looking to see what is open before proposing as your change proposal came in at least 10 - 15 minutes after my original one with v0.38.0 which is why I went with mine.

Patch-set: 4