Update patch set 1
Patch Set 1: Code-Review+2

> Patch Set 1:
> 
> Triggering on a daily basis if there are no new changes to the images will definitely fail these jobs if we are planning to use the releases push registry. 
> 
> Do we want to consume resources even if there are no new changes added to the images?

Jessica, the default is no cron. If the merge is used as a direct release vehicle, then yes, in our registry setup it would fail. It might not in another registry like DockerHub.

If it's used for RC builds, then it's not going to fail in our registry since we allow re-deploy on the 'staging' registry.

So, with all of that in mind, I'm going to go ahead and take this change as it's just adding flexibility that projects may desire and since it doesn't add a default cron, I think it's probably a good idea.

Patch-set: 1
Label: Code-Review=+2