Update patch set 1
Patch Set 1:

Thanks for the merge.

A cron-based trigger allows a Docker image to keep up with upstream changes. For example as a base "From" Debian/Stretch image evolves to incorporate security patches etc., a project can keep their specific image up-to-date as possible with that, and not grow stale.

Patch-set: 1