Update patch set 3
Patch Set 3:

(1 comment)

Chris. I think making the location of the yaml default to "DOCKER_ROOT" will be ok if we rely on the default DOCKER_ROOT value. But the script will break if for example DOCKER_ROOT is defined as "docker" (for example)
Then the script will cd into "docker" and will still try to look for the file in the "docker/location" path which will not work. 

I think the current "." value is the best since it is simple and we avoid not having to add much evaluation around it to predict the issues that will come up from what the user provides. 

Let me know your thoughts.

Patch-set: 3