Update patch set 4
Patch Set 4:

First let me say again that I would be happy to have this descoped down to just a change in the documentation, explaining that the dockerfile and yaml file must be both be in DOCKER_ROOT directory, which defaults to project repo root.

But if you are planning to continue with allowing a separate specification of the yaml file path, then it seems to me you'll probably want to modify the get-tag-yaml script so it no longer does "cd $DOCKER_ROOT" before getting to work -- instead just uses a path in a variable, which of course could then be defaulted to $DOCKER_ROOT.

Would it be better all around if I just withdraw my request (Jira issue)?  I didn't mean to create unnecessary work.

Patch-set: 4