Update patch set 4
Patch Set 4:

thanks for your comments Chris, I think is good that we are identifying places of improvement and if you see the need for this change (also something other projects can benefit from) then we should go ahead and continue to work on this patch better. 
Would it work if, as you say, we disable the cd $WORKSPACE in the tag scrip and just leave it to the build script to do it?

Patch-set: 4