Update patch set 4
Patch Set 4:

As I've been creating Jenkins jobs for the o-ran-sc projects I've hit various conflicting requirements for CI -- wow there's a surprise :)  The classic Docker projects that need to build an image are fine with a Dockerfile in an appropriate root directory, and adding one more tag file seems to be no big deal.  But some projects use a docker-based build only to verify changes (don't need to publish an image), and would prefer to tuck away metadata like Dockerfile and container-tag.yaml in a subdirectory.

I guess the obvious conclusion is that one size does not fit all, and the CI system should be flexible. So let's figure out a reasonable path here that accommodates various needs, and finish it up.

Patch-set: 4