Update patch set 4
Patch Set 4:

(1 comment)

I'm preparing an alternate version of this, hope to submit a new change set soon for you to consider.

One thing stumps me.  I would like to default the YAML directory to the value of the docker-root parameter, but cannot figure out how to make that be interpolated properly into the Jenkins config. It just appears as {docker-root}.

Patch-set: 4